requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.1
